+++ identity_stronghold/src/lib.rs
@@ -3,8 +3,11 @@
 
 pub(crate) mod ed25519;
 mod stronghold_jwk_storage;
+mod stronghold_jwk_storage_ext;
 mod stronghold_key_id;
+pub(crate) mod stronghold_key_type;
 #[cfg(test)]
 mod tests;
+pub(crate) mod utils;
 
 pub use stronghold_jwk_storage::*;
